Init commit
This is the initial commit that contains all the files as I had sent them to SJU. There are many issues here that I dont fully remember. I think there is a chance that I broke something and then never fixed it before sending and this doesnt work even a little bit.