Finished VideoCreator Script
I created the videocreator script, also compiled the project to an executable for the first time to test. I'm beginning to wonder if it would be a good idea to just combine these scipts into one script rather than a few tiny scripts, might refactor this later. I will probably continue on following what I did in python for now.